requirements.txt
pywebview
numpy
pandas
ebooklib
bs4
python-dotenv
deep_translator
spacy